% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audioplayRmd.R
\name{include_audio}
\alias{include_audio}
\title{Include audio}
\usage{
include_audio(audio_path, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{audio_path}{string with the path to the audio file (local or url)}

\item{width}{width (see \link[htmlwidgets:createWidget]{htmlwidgets::createWidget})}

\item{height}{height (see \link[htmlwidgets:createWidget]{htmlwidgets::createWidget})}

\item{elementId}{elementId (see \link[htmlwidgets:createWidget]{htmlwidgets::createWidget})}
}
\description{
Includes an audio player for rmarkdown and quarto.
It is the knitr::include_graphics() for audio files.
}
\examples{
include_audio("https://www2.cs.uic.edu/~i101/SoundFiles/PinkPanther30.wav")

}
